+++ src/CosmosDB/CosmosDB.Test/ScenarioTests/GremlinOperationsTests.cs
@@ -50,5 +50,19 @@ namespace Microsoft.Azure.Commands.CosmosDB.Test.ScenarioTests.ScenarioTest
         {
             TestRunner.RunTestScript("Test-GremlinMigrateThroughputCmdlets");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestGremlinInAccountRestoreOperationsCmdlets()
+        {
+            TestRunner.RunTestScript("Test-GremlinInAccountRestoreOperationsCmdlets");
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestGremlinInAccountRestoreOperationsSharedRUResourcesCmdlets()
+        {
+            TestRunner.RunTestScript("Test-GremlinInAccountRestoreOperationsSharedRUResourcesCmdlets");
+        }
     }
 }